% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utils.R
\name{colhex2col}
\alias{colhex2col}
\title{Convert a hex color to X11}
\usage{
colhex2col(colhex)
}
\arguments{
\item{colhex}{A hex color string}
}
\value{
returns the most similar X11 colour string
}
\description{
Convert a hex color to X11
}
\examples{
colhex2col("#A37BF0")
}
